# core: pass vm.Config by reference