--- a/eth/api_backend.go
+++ b/eth/api_backend.go
@@ -195,8 +195,7 @@ func (b *EthAPIBackend) GetEVM(ctx context.Context, msg core.Message, state *sta
 	vmError := func() error { return nil }
 
 	context := core.NewEVMContext(msg, header, b.eth.BlockChain(), nil)
-	config := *b.eth.blockchain.GetVMConfig()
-	return vm.NewEVM(context, state, b.eth.blockchain.Config(), &config), vmError, nil
+	return vm.NewEVM(context, state, b.eth.blockchain.Config(), b.eth.blockchain.GetVMConfig()), vmError, nil
 }
 
 func (b *EthAPIBackend) SubscribeRemovedLogsEvent(ch chan<- core.RemovedLogsEvent) event.Subscription {