The `vm.Config` object is ~2136 bytes in size so it makes sense to pass it by reference not by value.
This greatly reduces the amount of bytes we need to allocate for every call.

Master:
```
BenchmarkInsertChain_ring1000_diskdb-16    	      79	  13206599 ns/op	 5242574 B/op	   38345 allocs/op
BenchmarkInsertChain_ring1000_memdb-16    	      75	  13402992 ns/op	 5164621 B/op	   39069 allocs/op
```
This branch:
```
BenchmarkInsertChain_ring1000_diskdb-16    	      98	  13090816 ns/op	 4709312 B/op	   38572 allocs/op
BenchmarkInsertChain_ring1000_memdb-16    	      91	  12833283 ns/op	 4700052 B/op	   39268 allocs/op
```