@MariusVanDerWijden 2020-09-14
![Screenshot_2020-09-14_18-32-42](https://user-images.githubusercontent.com/16664698/93112945-dd927600-f6b8-11ea-9d0b-35c803be69ca.png)

Bench03: this pr, bench04: master
It seems that this PR is slightly slower but within ~2% of master. I'm not sure if this is because of differences in machines/peers or raw performance. Since the synthetic benchmarks suggest a slight speedup I think we could still go for it