@adamschmideg 2020-10-06
- [ ] Add nil check