@fjl 2020-10-06
Please make it possible to pass a nil `*vm.Config` to `core.NewBlockChain`. Alternatively, since we're changing the constructor already, maybe we can figure out a way to remove this parameter.