@adamschmideg 2020-10-20
An alternative solution is #21672